Add support for a pre-smart-pull hook
User can create and executable script in
.git/hooks/smart-pull-pre-fetch that will
be executed just before a fetch.

Example: There are tags in remote that need to be
cleaned out. One team member cleans them out but
then they come back the next time someone else
pushes a local tag before syncing local tags to
remote. A solution to this problem is to create
a smart-pull-pre-fetch hook that syncs local
tags to remote tags, so that old tags previously
deleted on remote don't come back.